Move get_request_parameter_length to libapi
